More clang warning fixes!

+++ src/control/controls.cpp
@@ -1267,7 +1267,7 @@ void RemoveNullInputMethods()
         if(Controls::g_InputMethods[i])
             i++;
         else
-            Controls::DeleteInputMethodSlot(i);
+            Controls::DeleteInputMethodSlot((int)i);
     }
 }
